Initial commit
Expects config.cfg in run directory with the same format as supplied
to Nethermind (can be the same file), where it extracts the "EthStats"
section to get the node name and websockets `eth-netstats` server connection info.

Expects to be running besides a Nethermind client so we can easily JsonRPC it on
`localhost:8545`.

Authenticates to `eth-netstats` and publishes the local Nethermind client's enode
on websockets channel `enode` once every minute.